feat: add shared schedule calendar component
